mcp-client-py/requirements.txt
mcp
python-dotenv 